requirements.txt
# RPS Gesture Referee System - Dependencies

# Core Dependencies
mediapipe==0.10.21
opencv-python==4.11.0.86
numpy==1.26.4
opencv-contrib-python==4.11.0.86
pyyaml>=6.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional Dependencies (Phase 4)
# playsound==1.2.2  # Audio support
# flask>=2.0.0      # Web interface
